+++ akamai/edgegrid/edgerc.py
@@ -15,15 +15,8 @@
 # limitations under the License.
 
 import logging
-import sys
+from configparser import ConfigParser
 from os.path import expanduser
-
-if sys.version_info[0] >= 3:
-    # python3
-    from configparser import ConfigParser
-else:
-    # python2.7
-    from ConfigParser import ConfigParser
 
 logger = logging.getLogger(__name__)