+++ HISTORY.rst
@@ -3,13 +3,13 @@
 History
 -------
 
-X.X.X (X-X-X)
+2.0.0 (2024-11-12)
 ++++++++++++++++++
 
 * Breaking changes
     - discontinue support for Python 2.7, the minimum supported version is now Python 3.9
     - the ``__init__`` function of ``EdgeGridAuth`` and ``EdgeGridAuthHeaders`` now accepts ``headers_to_sign`` and ``max_body`` as keyword-only arguments
-    - changed signatures of following methods in the ``EdgeGridAuthHeaders`` class: ``make_auth_header``, ``sign_request`` and ``make_data_to_sign``
+    - change signatures of these methods in the ``EdgeGridAuthHeaders`` class: ``make_auth_header``, ``sign_request`` and ``make_data_to_sign``
 
 * Improvements
     - update several dependencies in ``setup.py``
@@ -17,7 +17,7 @@ X.X.X (X-X-X)
       so that they contain the full set of project dependencies
 
 * Bug fixes
-    - properly handle file objects in the request body in EdgeGridAuth
+    - fix handling file objects in the request body for ``EdgeGridAuth``
 
 1.3.1 (2022-09-22)
 ++++++++++++++++++